Add Slider story, slider is now independant!
