+++ app/scripts/components/sliders.components.jsx
@@ -4,9 +4,10 @@ import Lifespan from 'lifespan';
 import PureRenderMixin from 'react-addons-pure-render-mixin';
 
 import LocalClient from '../stores/local-client.stores.jsx';
-import DOM from '../helpers/dom.helpers.js';
 import {indivGroupsEditionTutorialLabel} from '../helpers/joyride.helpers.js';
 import SliderHelpText from '../../images/sliders/helpText.json';
+
+import SliderController from './slider-controller.components';
 
 const demoRatio = 0.2;
 
@@ -30,7 +31,20 @@ export class Sliders extends React.PureComponent {
 				});
 			})
 			.onDelete(() => {
-				this.setState(undefined);
+				this.setState({values: undefined});
+			});
+
+		this.client.getStore('/userStore', this.lifespan)
+			.onUpdate((head) => {
+				// TODO: this should be changed since subscription would not mean pro anymore
+				const {subscription} = head.toJS().d;
+
+				if (subscription) {
+					this.setState({plan: subscription.plan.id});
+				}
+			})
+			.onDelete(() => {
+				this.setState({plan: undefined});
 			});
 		this.client.getStore('/userStore', this.lifespan)
 			.onUpdate((head) => {
@@ -138,6 +152,7 @@ export class Slider extends React.Component {
 		super(props);
 		this.shouldComponentUpdate = PureRenderMixin.shouldComponentUpdate.bind(this);
 		this.changeParam = this.changeParam.bind(this);
+		this.handleRestrictedRangeEnter = this.handleRestrictedRangeEnter.bind(this);
 	}
 
 	componentWillMount() {
@@ -151,6 +166,13 @@ export class Slider extends React.Component {
 
 	resetValue() {
 		this.client.dispatchAction('/change-param', {value: this.props.init, name: this.props.name, label: this.props.label, demo: this.props.demo});
+	}
+
+	handleRestrictedRangeEnter() {
+		this.client.dispatchAction('/store-value', {
+			openRestrictedFeature: true,
+			restrictedFeatureHovered: 'slider',
+		});
 	}
 
 	showTooltip(sliderName) {
@@ -254,7 +276,9 @@ export class Slider extends React.Component {
 						maxAdvised={this.props.maxAdvised}
 						disabled={this.props.disabled}
 						changeParam={this.changeParam}
-						child={this.props.child}/>
+						child={this.props.child}
+						onRestrictedRangeEnter={this.handleRestrictedRangeEnter}
+					/>
 					{indivSwitch}
 				</div>
 			</div>
@@ -321,162 +345,6 @@ export class RadioSlider extends React.Component {
 						radioValues={this.props.radioValues}/>
 					{indivSwitch}
 				</div>
-			</div>
-		);
-	}
-}
-
-export class SliderController extends React.Component {
-	constructor(props) {
-		super(props);
-		this.shouldComponentUpdate = PureRenderMixin.shouldComponentUpdate.bind(this);
-
-		// function bindings
-		this.handleUp = this.handleUp.bind(this);
-		this.handleMove = this.handleMove.bind(this);
-		this.handleSelectstart = this.handleSelectstart.bind(this);
-		this.restrictedRangeEnter = this.restrictedRangeEnter.bind(this);
-	}
-
-	componentWillMount() {
-		this.lifespan = new Lifespan();
-		this.client = LocalClient.instance();
-
-		document.addEventListener('mouseup', this.handleUp);
-		window.addEventListener('mousemove', this.handleMove);
-		document.addEventListener('selectstart', this.handleSelectstart);
-	}
-
-	componentDidMount() {
-		const slider = this.refs.slider;
-
-		this.sliderWidth = slider.offsetWidth;
-	}
-
-	componentWillUnmount() {
-		this.lifespan.release();
-		document.removeEventListener('mouseup', this.handleUp);
-		window.removeEventListener('mousemove', this.handleMove);
-		document.removeEventListener('selectstart', this.handleSelectstart);
-	}
-
-	handleDown(e) {
-		this.tracking = true;
-		const newX = e.pageX || e.screenX;
-		const {offsetLeft} = DOM.getAbsOffset(this.refs.slider);
-		let newValue = ((newX - offsetLeft) / this.sliderWidth * (this.props.max - this.props.min)) + this.props.min;
-
-		newValue = Math.min(Math.max(newValue, this.props.realMin), this.props.realMax);
-
-		this.props.changeParam({value: newValue, name: this.props.name, label: this.props.label});
-		this.currentX = newX;
-
-		e.stopPropagation();
-	}
-
-	handleUp(e) {
-		if (!this.tracking) {
-			return;
-		}
-
-		this.tracking = false;
-		this.props.changeParam({value: this.props.value, name: this.props.name, label: this.props.label, force: true});
-
-		e.stopPropagation();
-	}
-
-	restrictedRangeEnter(e) {
-		this.client.dispatchAction('/store-value', {openRestrictedFeature: true,
-													restrictedFeatureHovered: 'slider'});
-		e.stopPropagation();
-	}
-
-	handleMove(e) {
-		if (!this.tracking) {
-			return;
-		}
-
-		const newX = e.pageX || e.screenX;
-		const el = this.refs.slider;
-		const {offsetLeft} = DOM.getAbsOffset(el);
-		let newValue;
-
-		if (newX >= offsetLeft && newX <= offsetLeft + el.clientWidth) {
-			const variation = (newX - this.currentX) / this.sliderWidth * (this.props.max - this.props.min);
-
-			newValue = this.props.value + variation;
-
-			newValue = Math.min(Math.max(newValue, this.props.realMin), this.props.realMax);
-		}
-		else {
-			newValue = newX < offsetLeft ? this.props.realMin : this.props.realMax;
-		}
-
-		this.props.changeParam({value: newValue, name: this.props.name});
-		this.currentX = newX;
-	}
-
-	// This prevents preview text to be selected whil using the sliders
-	handleSelectstart(e) {
-		if (this.tracking) {
-			return e.preventDefault();
-		}
-	}
-
-	handleClick() {
-
-	}
-
-	render() {
-		const translateX = (this.props.max - Math.min(Math.max(this.props.value, this.props.min), this.props.max)) / (this.props.max - this.props.min) * 92.0;
-		const translateDemoMin = (this.props.max - this.props.realMin) / (this.props.max - this.props.min) * 92.0;
-		const translateDemoMax = 100 - ((this.props.max - this.props.realMax) / (this.props.max - this.props.min) * 92.0);
-		const transform = {
-			transform: `translateX(-${translateX}%)`,
-		};
-		const transformDemoMin = {
-			transform: `translateX(-${translateDemoMin}%)`,
-			marginLeft: '-10px',
-		};
-		const transformDemoMax = {
-			transform: `translateX(${translateDemoMax}%)`,
-			marginLeft: '-10px',
-		};
-
-		const classes = classNames({
-			'slider-controller-bg': true,
-			'is-not-advised': this.props.value < this.props.minAdvised || this.props.value > this.props.maxAdvised,
-			'is-indiv': this.props.individualized,
-		});
-
-		const demoClassesMin = classNames({
-			'slider-range-limiter-bg': true,
-			'min': true,
-		});
-
-		const demoClassesMax = classNames({
-			'slider-range-limiter-bg': true,
-			'max': true,
-		});
-
-		const demoRangeLimiters = this.props.demo
-			? (
-				<span>
-					<div onClick={this.restrictedRangeEnter} className={demoClassesMin} style={transformDemoMin}/>
-					<div onClick={this.restrictedRangeEnter} className={demoClassesMax} style={transformDemoMax}/>
-				</span>
-			)
-			: false;
-
-		return (
-			<div className="slider-controller" ref="slider"
-				onMouseDown={(e) => { this.handleDown(e);}}>
-				<div className={classes} style={transform}>
-					<div
-						className="slider-controller-handle"
-						ref="handle" ></div>
-				</div>
-				{demoRangeLimiters}
 			</div>
 		);
 	}
