+++ stories/index.js
@@ -1,4 +1,5 @@
 import '../app/scripts/styles';
 
+import './SliderController.story';
 import './AccountManageSubUsers.story';
 import './AccountValidationButton.story';
